Update tags_manager.php
Replaced the rest of the ternaries with more human-parseable code. We really don't care about extremely trivial "efficiency" in this web app, but about phpBB admins' ability to adapt the code. Also fixed and moved a side comment (no longer in vague wording, nor confusing a function with a variable). More code consistency. Fixed misuse of ".." for "..." as an indicator of continuation (".." doesn't serve that function in English). And injecting a line break between `if` and its `{` isn't objectively helpful while making the script page much longer to scroll through for no practical reason, so don't do it. (I left that alone with `function ... {` because it's done with a line break consistently throughout this entire codebase, and is not a big deal.)